+++ src/blocked/benchmark.cpp
@@ -9,10 +9,10 @@
 #include <vector>
 
 // Function prototype for blocked MMul
-void blocked_mmul(const double *A, const double *B, double *C, std::size_t N);
+void blocked_mmul(const float *A, const float *B, float *C, std::size_t N);
 
 // Function prototype for blocked parallel MMul
-void blocked_parallel_mmul(const double *A, const double *b, double *C,
+void blocked_parallel_mmul(const float *A, const float *b, float *C,
                            std::size_t N, std::size_t start_row,
                            std::size_t end_row);
 
@@ -24,12 +24,12 @@ static void blocked_mmul_bench(benchmark::State &s) {
   // Create our random number generators
   std::mt19937 rng;
   rng.seed(std::random_device()());
-  std::uniform_real_distribution<double> dist(-10, 10);
+  std::uniform_real_distribution<float> dist(-10, 10);
 
   // Create input matrices
-  double *A = new double[N * N];
-  double *B = new double[N * N];
-  double *C = new double[N * N];
+  float *A = new float[N * N];
+  float *B = new float[N * N];
+  float *C = new float[N * N];
 
   // Initialize them with random values (and C to 0)
   std::generate(A, A + N * N, [&] { return dist(rng); });
@@ -60,12 +60,12 @@ static void blocked_aligned_mmul_bench(benchmark::State &s) {
   // Create our random number generators
   std::mt19937 rng;
   rng.seed(std::random_device()());
-  std::uniform_real_distribution<double> dist(-10, 10);
+  std::uniform_real_distribution<float> dist(-10, 10);
 
   // Create input matrices
-  double *A = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *B = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *C = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
+  float *A = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *B = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *C = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
 
   // Initialize them with random values (and C to 0)
   std::generate(A, A + N * N, [&] { return dist(rng); });
@@ -96,12 +96,12 @@ static void parallel_blocked_mmul_bench(benchmark::State &s) {
   // Create our random number generators
   std::mt19937 rng;
   rng.seed(std::random_device()());
-  std::uniform_real_distribution<double> dist(-10, 10);
+  std::uniform_real_distribution<float> dist(-10, 10);
 
   // Create input matrices
-  double *A = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *B = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *C = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
+  float *A = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *B = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *C = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
 
   // Initialize them with random values (and C to 0)
   std::generate(A, A + N * N, [&] { return dist(rng); });
