+++ src/blocked/blocked_parallel_mmul.cpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 
 // Blocked parallel implementation
-void blocked_parallel_mmul(const double *A, const double *B, double *C,
+void blocked_parallel_mmul(const float *A, const float *B, float *C,
                            std::size_t N, std::size_t start_row,
                            std::size_t end_row) {
   // For each row...
