+++ src/blocked_column/blocked_column_mmul.cpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 
 // Blocked column serial implementation
-void blocked_column_mmul(const double *A, const double *B, double *C,
+void blocked_column_mmul(const float *A, const float *B, float *C,
                          std::size_t N) {
   // For each chunk of columns
   for (std::size_t col_chunk = 0; col_chunk < N; col_chunk += 16)
