+++ src/blocked_column_multi_output/benchmark.cpp
@@ -9,12 +9,12 @@
 #include <vector>
 
 // Function prototype for blocked column multi-output serial MMul
-void blocked_column_multi_output_mmul(const double *A, const double *B,
-                                      double *C, std::size_t N);
+void blocked_column_multi_output_mmul(const float *A, const float *B,
+                                      float *C, std::size_t N);
 
 // Function prototype for blocked column multi-output parallel MMul
-void blocked_column_multi_output_parallel_mmul(const double *A, const double *B,
-                                               double *C, std::size_t N,
+void blocked_column_multi_output_parallel_mmul(const float *A, const float *B,
+                                               float *C, std::size_t N,
                                                std::size_t start_col,
                                                std::size_t end_col);
 
@@ -27,12 +27,12 @@ static void blocked_column_multi_output_aligned_mmul_bench(
   // Create our random number generators
   std::mt19937 rng;
   rng.seed(std::random_device()());
-  std::uniform_real_distribution<double> dist(-10, 10);
+  std::uniform_real_distribution<float> dist(-10, 10);
 
   // Create input matrices
-  double *A = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *B = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *C = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
+  float *A = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *B = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *C = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
 
   // Initialize them with random values (and C to 0)
   std::generate(A, A + N * N, [&] { return dist(rng); });
@@ -64,12 +64,12 @@ static void parallel_blocked_column_multi_output_mmul_bench(
   // Create our random number generators
   std::mt19937 rng;
   rng.seed(std::random_device()());
-  std::uniform_real_distribution<double> dist(-10, 10);
+  std::uniform_real_distribution<float> dist(-10, 10);
 
   // Create input matrices
-  double *A = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *B = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
-  double *C = static_cast<double *>(aligned_alloc(64, N * N * sizeof(double)));
+  float *A = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *B = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
+  float *C = static_cast<float *>(aligned_alloc(64, N * N * sizeof(float)));
 
   // Initialize them with random values (and C to 0)
   std::generate(A, A + N * N, [&] { return dist(rng); });
